+++ pkg/functions/vshn-postgres/url/command.go
@@ -1,10 +1,10 @@
-package main
+package url
 
 import (
 	"github.com/go-logr/logr"
 	"github.com/urfave/cli/v2"
-	"github.com/vshn/appcat-comp-functions/functions"
-	"github.com/vshn/appcat-comp-functions/src"
+	"github.com/vshn/appcat-comp-functions/pkg"
+	"github.com/vshn/appcat-comp-functions/pkg/functions/vshn-postgres"
 )
 
 type vshnPostgresURL struct{}
@@ -12,15 +12,15 @@ type vshnPostgresURL struct{}
 func NewVshnPostgresURL() *cli.Command {
 	command := &vshnPostgresURL{}
 	return &cli.Command{
-		Name:   "vshn-postgres-url-function-io",
+		Name:   "vshn-postgres-url",
 		Usage:  "Start VSHN Postgres URL Function IO",
 		Action: command.execute,
 	}
 }
 
 func (c *vshnPostgresURL) execute(ctx *cli.Context) error {
-	_ = functions.LogMetadata(ctx, A)
+	_ = pkg.LogMetadata(ctx, vshnpostgres.AI)
 	log := logr.FromContextOrDiscard(ctx.Context).WithName(ctx.Command.Name)
 	log.Info("Executing FunctionIO - VSHN Postgres URL", "config", c)
-	return src.Exec(ctx.Context, transform)
+	return pkg.Exec(ctx.Context, transform)
 }
