+++ pkg/logger.go
@@ -1,4 +1,4 @@
-package functions
+package pkg
 
 import (
 	"fmt"
@@ -13,8 +13,9 @@ import (
 	"go.uber.org/zap/zapcore"
 )
 
-type AppInfo struct {
-	Version, Commit, Date, AppName, AppLongName string
+func init() {
+	// Remove `-v` short option from --version flag in favor of verbosity.
+	cli.VersionFlag.(*cli.BoolFlag).Aliases = nil
 }
 
 // LogMetadata prints various metadata to the root logger.
@@ -65,4 +66,27 @@ func newZapLogger(name, version string, verbosityLevel int, useProductionConfig 
 		return zlog.WithValues("version", version), nil
 	}
 	return zlog, nil
+}
+
+func NewLogLevelFlag() *cli.IntFlag {
+	return &cli.IntFlag{
+		Name: "log-level", Aliases: []string{"v"}, EnvVars: []string{"LOG_LEVEL"},
+		Usage: "number of the log level verbosity",
+		Value: 0,
+	}
+}
+
+func NewLogFormatFlag() *cli.StringFlag {
+	return &cli.StringFlag{
+		Name: "log-format", EnvVars: []string{"LOG_FORMAT"},
+		Usage: "sets the log format (values: [json, console])",
+		Value: "console",
+		Action: func(context *cli.Context, format string) error {
+			if format == "console" || format == "json" {
+				return nil
+			}
+			_ = cli.ShowAppHelp(context)
+			return fmt.Errorf("unknown log format: %s", format)
+		},
+	}
 }
