+++ pkg/types.go
@@ -1,0 +1,6 @@
+package pkg
+
+// AppInfo defines application information
+type AppInfo struct {
+	Version, Commit, Date, AppName, AppLongName string
+}
